+++ include/event/header_event.h
@@ -5,11 +5,11 @@
 
 class HeaderEvent : public Event {
 public:
-  const uint64_t from;
-  const uint64_t to;
+  const uint16_t from;
+  const uint16_t to;
   const std::shared_ptr<Header> header;
 
-  HeaderEvent(uint64_t timestamp, uint64_t from, uint64_t to, const std::shared_ptr<Header> &header);
+  HeaderEvent(uint64_t timestamp, uint16_t from, uint16_t to, const std::shared_ptr<Header> &header);
 
   void process(std::priority_queue<Event *, std::vector<Event *>, CompareEvent> &queue,
                const std::vector<std::unique_ptr<Node>> &nodes) const;
