+++ src/event/block_hash_event.cpp
@@ -1,7 +1,7 @@
 #include "event/block_hash_event.h"
 #include "emu/config.h"
 
-BlockHashEvent::BlockHashEvent(uint64_t timestamp, uint64_t from, uint64_t to, uint64_t blockHash)
+BlockHashEvent::BlockHashEvent(uint64_t timestamp, uint16_t from, uint16_t to, uint64_t blockHash)
     : Event(timestamp), from(from), to(to), blockHash(blockHash) {}
 
 void BlockHashEvent::process(std::priority_queue<Event *, std::vector<Event *>, CompareEvent> &queue,
