+++ src/event/body_fetch_timer_event.cpp
@@ -4,7 +4,7 @@
 
 const uint64_t bodyFetchInterval = 100;
 
-BodyFetchTimerEvent::BodyFetchTimerEvent(uint64_t timestamp, uint64_t id) : Event(timestamp), id(id) {}
+BodyFetchTimerEvent::BodyFetchTimerEvent(uint64_t timestamp, uint16_t id) : Event(timestamp), id(id) {}
 
 void BodyFetchTimerEvent::process(std::priority_queue<Event *, std::vector<Event *>, CompareEvent> &queue,
                                   const std::vector<std::unique_ptr<Node>> &nodes) const {
@@ -14,7 +14,7 @@ void BodyFetchTimerEvent::process(std::priority_queue<Event *, std::vector<Event
     return;
   }
   for (auto &[blockHash, owners] : node->fetchingBodies) {
-    uint64_t to = owners[rand() % owners.size()];
+    uint16_t to = owners[rand() % owners.size()];
     uint64_t interval = global.minDelay + rand() % (global.maxDelay - global.minDelay);
     queue.push(new BodyReqEvent(timestamp + interval, id, to, blockHash));
   }
